Cluster Estimated Standard Errors

Cluster Estimated Standard Errors (CESE)

vcovCESE(mod, cluster = NULL, type = NULL)

Arguments

mod
a model object. It can be the output of the functions lm, glm, or other regression function that returns compatible objects.

cluster
either a string vector with the name of the variables that will be used to cluster the standard errors, or a formula - e.g., ~ rhs, with a summation of the variables that will be used to cluster the standard errors replacing the rhs -, or a vector, matrix, or data.frame with the clustering data.

type
string with either HC0, HC1, HC2, HC3, or HC4. It specifies the type of heteroskedasticity correction to use (see Davidson and MacKinnon (1993) and Hayes and Cai (2007)).

Value

The function returns a variance-covariace matrix of the coefficient estimates using the Cluster Estimated Standard Error (CESE) method.

References

Jackson, John (2019) Corrected Standard Errors with Clustered Data. Political Analysis.

Hayes, A. F., & Cai, L., (2007) Using heteroskedasticity-consistent standard error estimators in ols regression: an introduction and software implementation, Behavior research methods, 39(4), 709–722.

Davidson, R., & MacKinnon, J. G., (2004) Econometric theory and methods: Oxford University Press New York.

Examples

mod  = lm(enep ~  enpc + fapres + enpcfapres + proximity + eneg + logmag + logmag_eneg , data=dcese)

## --------------------------------------
## Getting the variance covariance matrix
## -------------------------------------- 
## Original variance-covariance matrix (no clustered std. errors)
vcov(mod)

#>             (Intercept)         enpc        fapres    enpcfapres     proximity
#> (Intercept)  0.34193367 -0.080109404 -6.498717e-02  0.0227605333 -0.0416369373
#> enpc        -0.08010940  0.035696550  2.401318e-02 -0.0102825206  0.0059204296
#> fapres      -0.06498717  0.024013177  2.734250e-02 -0.0090017922 -0.0004345205
#> enpcfapres   0.02276053 -0.010282521 -9.001792e-03  0.0036429747 -0.0014388263
#> proximity   -0.04163694  0.005920430 -4.345205e-04 -0.0014388263  0.0776196264
#> eneg        -0.03580496 -0.001476759 -2.517854e-03  0.0007025382 -0.0039084482
#> logmag      -0.05448255 -0.006980873  1.742043e-04  0.0021399533 -0.0023836036
#> logmag_eneg  0.02532042  0.001832590 -7.512686e-05 -0.0007721436 -0.0009085951
#>                      eneg        logmag   logmag_eneg
#> (Intercept) -0.0358049582 -0.0544825503  2.532042e-02
#> enpc        -0.0014767587 -0.0069808731  1.832590e-03
#> fapres      -0.0025178544  0.0001742043 -7.512686e-05
#> enpcfapres   0.0007025382  0.0021399533 -7.721436e-04
#> proximity   -0.0039084482 -0.0023836036 -9.085951e-04
#> eneg         0.0218856408  0.0222887158 -1.190289e-02
#> logmag       0.0222887158  0.0606796325 -2.995518e-02
#> logmag_eneg -0.0119028856 -0.0299551838  1.778317e-02

## Variance-covariance matrix using CRSE (sandwish package)
## sandwich::vcovCL(mod, cluster = ~ country)
## sandwich::vcovCL(mod, cluster = ~ country, type="HC3")

## Variance-covariance matrix using CESE
ceser::vcovCESE(mod, cluster = ~ country)

#>             (Intercept)          enpc       fapres   enpcfapres    proximity
#> (Intercept)  1.51221650 -0.3374255708 -0.308548972  0.087864892 -0.082214926
#> enpc        -0.33742557  0.1187065304  0.099198869 -0.033556819 -0.003174936
#> fapres      -0.30854897  0.0991988687  0.135525330 -0.036883799 -0.016929595
#> enpcfapres   0.08786489 -0.0335568193 -0.036883799  0.012012890  0.003051838
#> proximity   -0.08221493 -0.0031749364 -0.016929595  0.003051838  0.132039888
#> eneg        -0.08262488  0.0026806954 -0.006697128  0.001033777 -0.005366609
#> logmag      -0.21216040  0.0009545731  0.006332777  0.003599288  0.009265360
#> logmag_eneg  0.07928461 -0.0055220287 -0.010890411  0.000813923  0.004237610
#>                     eneg        logmag  logmag_eneg
#> (Intercept) -0.082624881 -0.2121603981  0.079284613
#> enpc         0.002680695  0.0009545731 -0.005522029
#> fapres      -0.006697128  0.0063327767 -0.010890411
#> enpcfapres   0.001033777  0.0035992881  0.000813923
#> proximity   -0.005366609  0.0092653603  0.004237610
#> eneg         0.037272375  0.0455315878 -0.021837494
#> logmag       0.045531588  0.2123302156 -0.099187086
#> logmag_eneg -0.021837494 -0.0991870859  0.057389251

ceser::vcovCESE(mod, cluster = ~ country, type="HC3") # HC3 correction

#>             (Intercept)          enpc       fapres    enpcfapres    proximity
#> (Intercept)  1.59804215 -0.3565889644 -0.326044522  0.0928614018 -0.086958688
#> enpc        -0.35658896  0.1254735451  0.104834222 -0.0354704134 -0.003332760
#> fapres      -0.32604452  0.1048342220  0.143205927 -0.0389793642 -0.017878997
#> enpcfapres   0.09286140 -0.0354704134 -0.038979364  0.0126978025  0.003217871
#> proximity   -0.08695869 -0.0033327597 -0.017878997  0.0032178709  0.139694748
#> eneg        -0.08737194  0.0028258190 -0.007080605  0.0010940285 -0.005680272
#> logmag      -0.22422347  0.0009844983  0.006687954  0.0038080340  0.009776123
#> logmag_eneg  0.08380931 -0.0058250477 -0.011500450  0.0008568949  0.004471824
#>                     eneg        logmag   logmag_eneg
#> (Intercept) -0.087371945 -0.2242234723  0.0838093142
#> enpc         0.002825819  0.0009844983 -0.0058250477
#> fapres      -0.007080605  0.0066879544 -0.0115004496
#> enpcfapres   0.001094029  0.0038080340  0.0008568949
#> proximity   -0.005680272  0.0097761226  0.0044718243
#> eneg         0.039433115  0.0481561426 -0.0231003489
#> logmag       0.048156143  0.2244236813 -0.1048418029
#> logmag_eneg -0.023100349 -0.1048418029  0.0606625677

## ---------
## Summaries
## ---------
## no robust SE 
summary(mod)

#> 
#> Call:
#> lm(formula = enep ~ enpc + fapres + enpcfapres + proximity + 
#>     eneg + logmag + logmag_eneg, data = dcese)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -3.5592 -0.8192 -0.3615  0.3769  9.0393 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  2.70431    0.58475   4.625 5.65e-06 ***
#> enpc         0.30399    0.18894   1.609 0.108711    
#> fapres      -0.61179    0.16536  -3.700 0.000258 ***
#> enpcfapres   0.20779    0.06036   3.443 0.000660 ***
#> proximity   -0.02243    0.27860  -0.081 0.935890    
#> eneg        -0.06566    0.14794  -0.444 0.657483    
#> logmag      -0.18146    0.24633  -0.737 0.461926    
#> logmag_eneg  0.36051    0.13335   2.703 0.007266 ** 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 1.653 on 291 degrees of freedom
#> Multiple R-squared:  0.3784,	Adjusted R-squared:  0.3634 
#> F-statistic: 25.31 on 7 and 291 DF,  p-value: < 2.2e-16
#> 

## summary table using CRSE (sandwich package)
## lmtest::coeftest(mod, vcov = sandwich::vcovCL, cluster = ~ country)                   

## summary using CESE
lmtest::coeftest(mod, vcov = ceser::vcovCESE, cluster = ~ country, type='HC3')

#> 
#> t test of coefficients:
#> 
#>              Estimate Std. Error t value Pr(>|t|)  
#> (Intercept)  2.704309   1.264137  2.1393  0.03325 *
#> enpc         0.303988   0.354222  0.8582  0.39150  
#> fapres      -0.611787   0.378426 -1.6167  0.10703  
#> enpcfapres   0.207791   0.112685  1.8440  0.06620 .
#> proximity   -0.022429   0.373758 -0.0600  0.95219  
#> eneg        -0.065662   0.198578 -0.3307  0.74114  
#> logmag      -0.181462   0.473734 -0.3830  0.70197  
#> logmag_eneg  0.360513   0.246298  1.4637  0.14435  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>